Split out some license-related types to an isolated file
