+++ src/internal/Licensing/LicensingTypes.ts
@@ -1,0 +1,16 @@
+/**
+ * @internal
+ */
+export enum LicenseState {
+    UNLICENSED,
+    LICENSED,
+    CAMERA_UNLICENSED,
+}
+
+/**
+ * @internal
+ */
+export interface LicenseChangeResult {
+    changeDetails: string;
+    succeeded: boolean;
+}
